Update cinnamon.css
When hovering the cursor over the date and clock at the panel (default is in the right bottom corner), the font turns to normal (default is bold) which results in shifting the icons right. To fix it the value "normal" of the property "font-weight" of the "panel.js/ -> #panel" section has been changed to "bold".